langgraph/pyproject.toml
[project]
name = "a2a-sample-agent-langgraph"
version = "0.1.0"
description = "Sample LangGraph currency agent with A2A Protocol"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk==0.2.8",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.3.18",
    "langchain-openai >=0.1.0",
    "pandas>=2.0.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
